Add db:migrate:preview

+++ lib/ez/schema_modifier.rb
@@ -4,28 +4,38 @@ module EZ
   # and applies any necessary db schema changes.
   class SchemaModifier
 
-    attr_reader :db, :spec
+    attr_reader :db, :spec, :dry_run
 
-    def initialize(model_spec, silent = false)
+    def initialize(model_spec, silent = false, dry_run = false)
       @silent = silent
+      @dry_run = dry_run
       @spec = model_spec
       connect_to_database
     end
 
-    def self.migrate(model_spec, silent = false)
-      self.new(model_spec, silent).migrate
+    def self.migrate(model_spec, silent = false, dry_run = false)
+      self.new(model_spec, silent, dry_run).migrate
     end
 
     def migrate
       @changed = false
 
+      if @dry_run
+        puts "Previewing... no changes will actually take place..."
+        puts
+      end
+
       add_missing_schema
       remove_dead_schema
 
       if @changed
-        update_schema_version
+        update_schema_version unless @dry_run
+        puts "\n(No changes were actually made)" if @dry_run
       else
-        puts "All tables are up-to-date."
+          puts "All tables are up-to-date."
+        if @dry_run
+          puts "\nNo changes would be made."
+        end
       end
 
       return @changed
@@ -77,22 +87,23 @@ module EZ
           if !assume_missing
             display_change "Adding new column '#{col_name}' as #{col_type} for model #{model_name}"
           end
-          db.add_column(table_name, col_name.to_sym, col_type.to_sym, default: col_default)
+          db.add_column(table_name, col_name.to_sym, col_type.to_sym, default: col_default)  unless @dry_run
           if data[:index]
             display_change "  (adding database index for '#{col_name}')"
-            db.add_index table_name, col_name.to_sym
+            db.add_index table_name, col_name.to_sym unless @dry_run
           end
         else
           if db_col.type != col_type
             display_change "Changing column type for #{col_name} to #{col_type} for model #{model_name}"
           end
 
+          # puts "#{table_name} #{col_name}: #{db_col.default} and #{col_default}"
           if db_col.default != col_default
-            display_change "Applying new default value #{col_default} for #{col_name} for model #{model_name}"
+            display_change "Applying new default value #{col_default || 'NULL'} for #{col_name} for model #{model_name}"
           end
 
           if (db_col.type != col_type) || (db_col.default != col_default)
-            db.change_column(table_name, col_name.to_sym, col_type.to_sym, default: col_default)
+            db.change_column(table_name, col_name.to_sym, col_type.to_sym, default: col_default)  unless @dry_run
           end
         end
       end
@@ -101,7 +112,7 @@ module EZ
     def add_model(model_name, columns)
       table_name = model_name.tableize
       display_change "Defining new table for model '#{model_name}'."
-      db.create_table table_name
+      db.create_table table_name  unless @dry_run
       add_missing_columns model_name, columns, true
       filename = "app/models/#{model_name.underscore}.rb"
       unless Rails.env.production? || File.exists?(filename)
@@ -125,13 +136,15 @@ module EZ
         if @spec.has_key?(model_name)
           db_columns = db.columns(table_name).map { |column| column.name.to_sym } - [:id, :created_at, :updated_at]
           spec_columns = @spec[model_name].keys.map(&:to_sym)
+          # puts spec_columns.inspect
           dead_columns = db_columns - spec_columns
+
 
           if dead_columns.any?
             dead_columns.each do |dead_column_name|
               display_change "Removing unused column '#{dead_column_name}' from model '#{model_name}'"
             end
-            db.remove_columns(table_name, *dead_columns)
+            db.remove_columns(table_name, *dead_columns)  unless @dry_run
           end
         end
       end
@@ -149,7 +162,7 @@ module EZ
       dead_tables.each do |table_name|
         model_name = table_name.classify
         display_change "Dropping table #{table_name}"
-        db.drop_table(table_name)
+        db.drop_table(table_name) unless @dry_run
         begin
           filename = "app/models/#{model_name.underscore}.rb"
           code = IO.read(filename)
@@ -157,16 +170,12 @@ module EZ
 
           if is_empty
             display_change "Deleting file #{filename}"
-            File.unlink(filename)
+            File.unlink(filename) unless @dry_run
           end
         rescue => e
           display_change "Could not delete old model #{model_name.underscore}.rb."
         end
       end
-    end
-
-    def silence_migration_output
-      ActiveRecord::Migration.verbose = false
     end
 
     def connect_to_database
